fixed all delete links and now all cukes passing

+++ app/helpers/application_helper.rb
@@ -20,7 +20,9 @@ module ApplicationHelper
         ret += '<em>'
         ret += item.updated_at.strftime("Last updated %B %d @ %H:%M by #{item.updated_by.puid}")
         ret += '</em>'
-        ret += link_to "Delete", admin_page_path(item), :method => :delete, :confirm => "Are you sure you want to delete the page #{item.title}", :class => "delete" if can? :destroy, item
+        if can? :destroy, item
+          ret += link_to "Delete", admin_page_path(item), :method => :delete, :confirm => "Are you sure you want to delete the page #{item.title}", :class => "delete" if can? :destroy, item
+        end
         ret += '</div>'
         ret += tree_ul(item.children, false, &block) if item.children.size > 0
         ret += '</li>'
